[e2e] examples から e2e test を一旦移動する
