+++ .github/workflows/e2e-test.yml
@@ -1,4 +1,4 @@
-name: examples-e2e-test
+name: e2e-test
 
 on:
   push:
@@ -17,7 +17,7 @@ env:
   TEST_SECRET_KEY: ${{ secrets.TEST_SECRET_KEY }}
 
 jobs:
-  e2e_ubuntu_test:
+  e2e_test_ubuntu:
     if: >-
       ${{
         contains(github.event.head_commit.message, '[e2e]') ||
@@ -28,10 +28,7 @@ jobs:
         python_version: ["3.10", "3.11", "3.12"]
         os: ["ubuntu-22.04", "ubuntu-24.04"]
     runs-on: ${{ matrix.os}}
-    defaults:
-      run:
-        working-directory: ./examples
-    timeout-minutes: 20
+    timeout-minutes: 30
     steps:
       - uses: actions/checkout@v4
       # libx11-dev は Ubuntu 24.04 の時に必要になる模様
@@ -47,12 +44,13 @@ jobs:
       - uses: eifinger/setup-rye@v4
       - run: rye pin ${{ matrix.python_version }}
       - run: rye sync
+      - run: rye run python run.py ${{ matrix.os }}_x86_64
       - run: rye run pytest tests/test_openh264.py -s
       - run: rye run pytest tests/test_messaging.py -s
       - run: rye run pytest tests/test_sendonly_recvonly.py -s
       - run: rye run pytest tests/test_vad.py -s
 
-  e2e_macos_test:
+  e2e_test_macos:
     if: >-
       ${{
         contains(github.event.head_commit.message, '[e2e]') ||
@@ -64,10 +62,7 @@ jobs:
         # macos-13 は test_macos.py が上手くテストが動かないのでスキップ
         os: ["macos-14"]
     runs-on: ${{ matrix.os }}
-    defaults:
-      run:
-        working-directory: ./examples
-    timeout-minutes: 20
+    timeout-minutes: 30
     steps:
       - uses: actions/checkout@v4
       - name: Download openh264
@@ -79,13 +74,14 @@ jobs:
       - uses: eifinger/setup-rye@v4
       - run: rye pin ${{ matrix.python_version }}
       - run: rye sync
+      - run: rye run python run.py macos_arm64
       - run: rye run pytest tests/test_macos.py -s
       - run: rye run pytest tests/test_openh264.py -s
       - run: rye run pytest tests/test_messaging.py -s
       - run: rye run pytest tests/test_sendonly_recvonly.py -s
       - run: rye run pytest tests/test_vad.py -s
 
-  e2e_windows_test:
+  e2e_test_windows:
     if: >-
       ${{
         contains(github.event.head_commit.message, '[e2e]') ||
@@ -95,10 +91,7 @@ jobs:
       matrix:
         python_version: ["3.10", "3.11", "3.12"]
     runs-on: windows-2022
-    defaults:
-      run:
-        working-directory: ./examples
-    timeout-minutes: 20
+    timeout-minutes: 30
     env:
       # Python を強制的に UTF-8 で利用するおまじない
       PYTHONUTF8: 1
@@ -115,12 +108,13 @@ jobs:
       - uses: eifinger/setup-rye@v4
       - run: rye pin ${{ matrix.python_version }}
       - run: rye sync
+      - run: rye run python run.py windows_x86_64
       - run: rye run pytest tests/test_messaging.py -s
       - run: rye run pytest tests/test_sendonly_recvonly.py -s
       - run: rye run pytest tests/test_vad.py -s
 
   slack_notify_succeeded:
-    needs: [e2e_ubuntu_test, e2e_macos_test, e2e_windows_test]
+    needs: [e2e_test_ubuntu, e2e_test_macos, e2e_test_windows]
     runs-on: ubuntu-latest
     if: success()
     steps:
@@ -134,7 +128,7 @@ jobs:
           SLACK_WEBHOOK: ${{ secrets.SLACK_WEBHOOK }}
 
   slack_notify_failed:
-    needs: [e2e_ubuntu_test, e2e_macos_test, e2e_windows_test]
+    needs: [e2e_test_ubuntu, e2e_test_macos, e2e_test_windows]
     runs-on: ubuntu-latest
     if: failure()
     steps:
