--- examples/.gitignore
@@ -1,7 +1,0 @@
-.venv
-.ruff_cache
-.mypy_cache
-*.pyc
-__pycache__
-.env*
-!.env.template
