+++ examples/pyproject.toml
@@ -7,22 +7,14 @@ dependencies = [
     "opencv-python~=4.10.0.84",
     "opencv-python-headless~=4.10.0.84",
     "sounddevice~=0.4.7",
-    "sora-sdk>=2024.4.0.dev0",
+    "sora-sdk>=2024.3.0",
     "mediapipe~=0.10.14",
     "python-dotenv>=1.0.1",
     "numpy>=2.0.1",
+    "pillow>=10.4.0",
 ]
 readme = "README.md"
 requires-python = ">= 3.10"
-
-[project.scripts]
-media_sendonly = "media.sendonly:sendonly"
-media_recvonly = "media.recvonly:recvonly"
-media_vad = "media.vad:vad"
-messaging_sendrecv = "messaging.sendrecv:sendrecv"
-messaging_sendonly = "messaging.sendonly:sendonly"
-messaging_recvonly = "messaging.recvonly:recvonly"
-hideface_sender = "ml.hideface_sender:hideface_sender"
 
 [build-system]
 requires = ["hatchling"]
@@ -30,13 +22,17 @@ build-backend = "hatchling.build"
 
 [tool.rye]
 managed = true
-dev-dependencies = ["ruff>=0.6", "mypy>=1.10.0", "pytest>=8.3"]
+dev-dependencies = [
+    "ruff>=0.6",
+    "mypy>=1.11.2",
+    "pytest>=8.3",
+]
 
 [tool.hatch.metadata]
 allow-direct-references = true
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/media", "src/messaging", "src/ml"]
+packages = ["src"]
 
 [tool.ruff]
 target-version = "py310"
