+++ examples/requirements-dev.lock
@@ -14,9 +14,9 @@ absl-py==2.1.0
     # via mediapipe
 attrs==24.2.0
     # via mediapipe
-cffi==1.17.0
+cffi==1.17.1
     # via sounddevice
-contourpy==1.2.1
+contourpy==1.3.0
     # via matplotlib
 cycler==0.12.1
     # via matplotlib
@@ -33,19 +33,19 @@ jax==0.4.31
 jaxlib==0.4.31
     # via jax
     # via mediapipe
-kiwisolver==1.4.5
+kiwisolver==1.4.7
     # via matplotlib
 matplotlib==3.9.2
     # via mediapipe
 mediapipe==0.10.14
     # via sora-sdk-samples
-ml-dtypes==0.4.0
+ml-dtypes==0.5.0
     # via jax
     # via jaxlib
-mypy==1.11.1
+mypy==1.11.2
 mypy-extensions==1.0.0
     # via mypy
-numpy==2.1.0
+numpy==2.1.1
     # via contourpy
     # via jax
     # via jaxlib
@@ -71,21 +71,22 @@ packaging==24.1
     # via pytest
 pillow==10.4.0
     # via matplotlib
+    # via sora-sdk-samples
 pluggy==1.5.0
     # via pytest
 protobuf==4.25.4
     # via mediapipe
 pycparser==2.22
     # via cffi
-pyparsing==3.1.2
+pyparsing==3.1.4
     # via matplotlib
-pytest==8.3.2
+pytest==8.3.3
 python-dateutil==2.9.0.post0
     # via matplotlib
 python-dotenv==1.0.1
     # via sora-sdk-samples
-ruff==0.6.1
-scipy==1.14.0
+ruff==0.6.5
+scipy==1.14.1
     # via jax
     # via jaxlib
 six==1.16.0
