+++ examples/src/shiguremaru.png
