+++ examples/src/vad.py
