--- examples/tests/test_vad.py
@@ -1,48 +1,0 @@
-import sys
-import time
-import uuid
-
-from media import Sendonly
-from media.vad import VAD
-
-
-def test_vad(setup):
-    signaling_urls = setup.get("signaling_urls")
-    channel_id_prefix = setup.get("channel_id_prefix")
-    metadata = setup.get("metadata")
-
-    channel_id = f"{channel_id_prefix}_{__name__}_{sys._getframe().f_code.co_name}_{uuid.uuid4()}"
-
-    sendonly = Sendonly(signaling_urls, channel_id, metadata=metadata)
-    sendonly.connect(fake_audio=True)
-
-    vad = VAD(signaling_urls, channel_id, metadata=metadata)
-    vad.connect()
-
-    time.sleep(5)
-
-    sendonly_stats = sendonly.get_stats()
-    vad_stats = vad.get_stats()
-
-    sendonly.disconnect()
-    vad.disconnect()
-
-    # codec が無かったら StopIteration 例外が上がる
-    sendonly_codec_stats = next(s for s in sendonly_stats if s.get("type") == "codec")
-    assert sendonly_codec_stats["mimeType"] == "audio/opus"
-
-    # outbound-rtp が無かったら StopIteration 例外が上がる
-    outbound_rtp_stats = next(s for s in sendonly_stats if s.get("type") == "outbound-rtp")
-    # audio には encoderImplementation が無い
-    assert outbound_rtp_stats["bytesSent"] > 0
-    assert outbound_rtp_stats["packetsSent"] > 0
-
-    # codec が無かったら StopIteration 例外が上がる
-    vad_codec_stats = next(s for s in vad_stats if s.get("type") == "codec")
-    assert vad_codec_stats["mimeType"] == "audio/opus"
-
-    # outbound-rtp が無かったら StopIteration 例外が上がる
-    inbound_rtp_stats = next(s for s in vad_stats if s.get("type") == "inbound-rtp")
-    # audio には decoderImplementation が無い
-    assert inbound_rtp_stats["bytesReceived"] > 0
-    assert inbound_rtp_stats["packetsReceived"] > 0
