+++ requirements.lock
@@ -10,3 +10,5 @@
 #   universal: false
 
 -e file:.
+pillow==10.4.0
+    # via sora-sdk
