+++ tests/test_sendonly_recvonly.py
@@ -2,17 +2,10 @@ import sys
 import time
 import uuid
 
-import pytest
-
-from media import Recvonly, Sendonly
-
-"""
-GitHub Actions で Video Toolbox を送受信で利用しようとするとエラーになるので、
-テストを sendonly のみに絞っている
-"""
+from client import Recvonly, Sendonly
 
 
-def test_macos_h264_sendonly(setup):
+def test_sendonly_recvonly_opus(setup):
     signaling_urls = setup.get("signaling_urls")
     channel_id_prefix = setup.get("channel_id_prefix")
     metadata = setup.get("metadata")
@@ -22,95 +15,16 @@ def test_macos_h264_sendonly(setup):
     sendonly = Sendonly(
         signaling_urls,
         channel_id,
-        audio=False,
-        video=True,
-        video_codec_type="H264",
+        audio=True,
+        video=False,
         metadata=metadata,
-        use_hwa=True,
     )
-    sendonly.connect(fake_video=True)
-
-    time.sleep(5)
-
-    assert sendonly.connect_message.get("channel_id") == channel_id
-    assert sendonly.connect_message.get("video").get("codec_type") == "H264"
-
-    sendonly_stats = sendonly.get_stats()
-
-    sendonly.disconnect()
-
-    # codec が無かったら StopIteration 例外が上がる
-    codec_stats = next(s for s in sendonly_stats if s.get("type") == "codec")
-    # H.264 が採用されているかどうか確認する
-    assert codec_stats["mimeType"] == "video/H264"
-
-    # outbound-rtp が無かったら StopIteration 例外が上がる
-    outbound_rtp_stats = next(s for s in sendonly_stats if s.get("type") == "outbound-rtp")
-    assert outbound_rtp_stats["encoderImplementation"] == "VideoToolbox"
-    assert outbound_rtp_stats["bytesSent"] > 0
-    assert outbound_rtp_stats["packetsSent"] > 0
-
-
-def test_macos_h265_sendonly(setup):
-    signaling_urls = setup.get("signaling_urls")
-    channel_id_prefix = setup.get("channel_id_prefix")
-    metadata = setup.get("metadata")
-
-    channel_id = f"{channel_id_prefix}_{__name__}_{sys._getframe().f_code.co_name}_{uuid.uuid4()}"
-
-    sendonly = Sendonly(
-        signaling_urls,
-        channel_id,
-        audio=False,
-        video=True,
-        video_codec_type="H265",
-        metadata=metadata,
-        use_hwa=True,
-    )
-    sendonly.connect(fake_video=True)
-
-    time.sleep(5)
-
-    sendonly_stats = sendonly.get_stats()
-
-    sendonly.disconnect()
-
-    # codec が無かったら StopIteration 例外が上がる
-    codec_stats = next(s for s in sendonly_stats if s.get("type") == "codec")
-    # H.264 が採用されているかどうか確認する
-    assert codec_stats["mimeType"] == "video/H265"
-
-    # outbound-rtp が無かったら StopIteration 例外が上がる
-    outbound_rtp_stats = next(s for s in sendonly_stats if s.get("type") == "outbound-rtp")
-    assert outbound_rtp_stats["encoderImplementation"] == "VideoToolbox"
-    assert outbound_rtp_stats["bytesSent"] > 0
-    assert outbound_rtp_stats["packetsSent"] > 0
-
-
-@pytest.mark.skip(reason="ローカルでは成功する")
-def test_macos_h264_sendonly_recvonly(setup):
-    signaling_urls = setup.get("signaling_urls")
-    channel_id_prefix = setup.get("channel_id_prefix")
-    metadata = setup.get("metadata")
-
-    channel_id = f"{channel_id_prefix}_{__name__}_{sys._getframe().f_code.co_name}_{uuid.uuid4()}"
-
-    sendonly = Sendonly(
-        signaling_urls,
-        channel_id,
-        audio=False,
-        video=True,
-        video_codec_type="H264",
-        metadata=metadata,
-        use_hwa=True,
-    )
-    sendonly.connect()
+    sendonly.connect(fake_audio=True)
 
     recvonly = Recvonly(
         signaling_urls,
         channel_id,
         metadata=metadata,
-        use_hwa=True,
     )
     recvonly.connect()
 
@@ -124,29 +38,26 @@ def test_macos_h264_sendonly_recvonly(setup):
 
     # codec が無かったら StopIteration 例外が上がる
     sendonly_codec_stats = next(s for s in sendonly_stats if s.get("type") == "codec")
-    # H.264 が採用されているかどうか確認する
-    assert sendonly_codec_stats["mimeType"] == "video/H264"
+    assert sendonly_codec_stats["mimeType"] == "audio/opus"
 
     # outbound-rtp が無かったら StopIteration 例外が上がる
     outbound_rtp_stats = next(s for s in sendonly_stats if s.get("type") == "outbound-rtp")
-    assert outbound_rtp_stats["encoderImplementation"] == "VideoToolbox"
+    # audio には encoderImplementation が無い
     assert outbound_rtp_stats["bytesSent"] > 0
     assert outbound_rtp_stats["packetsSent"] > 0
 
     # codec が無かったら StopIteration 例外が上がる
     recvonly_codec_stats = next(s for s in recvonly_stats if s.get("type") == "codec")
-    # H.264 が採用されているかどうか確認する
-    assert recvonly_codec_stats["mimeType"] == "video/H264"
+    assert recvonly_codec_stats["mimeType"] == "audio/opus"
 
     # outbound-rtp が無かったら StopIteration 例外が上がる
     inbound_rtp_stats = next(s for s in recvonly_stats if s.get("type") == "inbound-rtp")
-    assert inbound_rtp_stats["decoderImplementation"] == "VideoToolbox"
+    # audio には decoderImplementation が無い
     assert inbound_rtp_stats["bytesReceived"] > 0
     assert inbound_rtp_stats["packetsReceived"] > 0
 
 
-@pytest.mark.skip(reason="ローカルでは成功する")
-def test_macos_h265_sendonly_recvonly(setup):
+def test_sendonly_recvonly_vp8(setup):
     signaling_urls = setup.get("signaling_urls")
     channel_id_prefix = setup.get("channel_id_prefix")
     metadata = setup.get("metadata")
@@ -158,17 +69,15 @@ def test_macos_h265_sendonly_recvonly(setup):
         channel_id,
         audio=False,
         video=True,
-        video_codec_type="H265",
+        video_codec_type="VP8",
         metadata=metadata,
-        use_hwa=True,
     )
-    sendonly.connect()
+    sendonly.connect(fake_video=True)
 
     recvonly = Recvonly(
         signaling_urls,
         channel_id,
         metadata=metadata,
-        use_hwa=True,
     )
     recvonly.connect()
 
@@ -182,20 +91,126 @@ def test_macos_h265_sendonly_recvonly(setup):
 
     # codec が無かったら StopIteration 例外が上がる
     sendonly_codec_stats = next(s for s in sendonly_stats if s.get("type") == "codec")
-    assert sendonly_codec_stats["mimeType"] == "video/H265"
+    assert sendonly_codec_stats["mimeType"] == "video/VP8"
 
     # outbound-rtp が無かったら StopIteration 例外が上がる
     outbound_rtp_stats = next(s for s in sendonly_stats if s.get("type") == "outbound-rtp")
-    assert outbound_rtp_stats["encoderImplementation"] == "VideoToolbox"
+    assert outbound_rtp_stats["encoderImplementation"] == "libvpx"
     assert outbound_rtp_stats["bytesSent"] > 0
     assert outbound_rtp_stats["packetsSent"] > 0
 
     # codec が無かったら StopIteration 例外が上がる
     recvonly_codec_stats = next(s for s in recvonly_stats if s.get("type") == "codec")
-    assert recvonly_codec_stats["mimeType"] == "video/H265"
+    assert recvonly_codec_stats["mimeType"] == "video/VP8"
 
     # outbound-rtp が無かったら StopIteration 例外が上がる
     inbound_rtp_stats = next(s for s in recvonly_stats if s.get("type") == "inbound-rtp")
-    assert inbound_rtp_stats["decoderImplementation"] == "VideoToolbox"
+    assert inbound_rtp_stats["decoderImplementation"] == "libvpx"
+    assert inbound_rtp_stats["bytesReceived"] > 0
+    assert inbound_rtp_stats["packetsReceived"] > 0
+
+
+def test_sendonly_recvonly_vp9(setup):
+    signaling_urls = setup.get("signaling_urls")
+    channel_id_prefix = setup.get("channel_id_prefix")
+    metadata = setup.get("metadata")
+
+    channel_id = f"{channel_id_prefix}_{__name__}_{sys._getframe().f_code.co_name}_{uuid.uuid4()}"
+
+    sendonly = Sendonly(
+        signaling_urls,
+        channel_id,
+        audio=False,
+        video=True,
+        video_codec_type="VP9",
+        metadata=metadata,
+    )
+    sendonly.connect(fake_video=True)
+
+    recvonly = Recvonly(
+        signaling_urls,
+        channel_id,
+        metadata,
+    )
+    recvonly.connect()
+
+    time.sleep(5)
+
+    sendonly_stats = sendonly.get_stats()
+    recvonly_stats = recvonly.get_stats()
+
+    sendonly.disconnect()
+    recvonly.disconnect()
+
+    # codec が無かったら StopIteration 例外が上がる
+    sendonly_codec_stats = next(s for s in sendonly_stats if s.get("type") == "codec")
+    assert sendonly_codec_stats["mimeType"] == "video/VP9"
+
+    # outbound-rtp が無かったら StopIteration 例外が上がる
+    outbound_rtp_stats = next(s for s in sendonly_stats if s.get("type") == "outbound-rtp")
+    assert outbound_rtp_stats["encoderImplementation"] == "libvpx"
+    assert outbound_rtp_stats["bytesSent"] > 0
+    assert outbound_rtp_stats["packetsSent"] > 0
+
+    # codec が無かったら StopIteration 例外が上がる
+    recvonly_codec_stats = next(s for s in recvonly_stats if s.get("type") == "codec")
+    assert recvonly_codec_stats["mimeType"] == "video/VP9"
+
+    # outbound-rtp が無かったら StopIteration 例外が上がる
+    inbound_rtp_stats = next(s for s in recvonly_stats if s.get("type") == "inbound-rtp")
+    assert inbound_rtp_stats["decoderImplementation"] == "libvpx"
+    assert inbound_rtp_stats["bytesReceived"] > 0
+    assert inbound_rtp_stats["packetsReceived"] > 0
+
+
+def test_sendonly_recvonly_av1(setup):
+    signaling_urls = setup.get("signaling_urls")
+    channel_id_prefix = setup.get("channel_id_prefix")
+    metadata = setup.get("metadata")
+
+    channel_id = f"{channel_id_prefix}_{__name__}_{sys._getframe().f_code.co_name}_{uuid.uuid4()}"
+
+    sendonly = Sendonly(
+        signaling_urls,
+        channel_id,
+        audio=False,
+        video=True,
+        video_codec_type="AV1",
+        metadata=metadata,
+    )
+    sendonly.connect(fake_video=True)
+
+    recvonly = Recvonly(
+        signaling_urls,
+        channel_id,
+        metadata=metadata,
+    )
+    recvonly.connect()
+
+    time.sleep(5)
+
+    sendonly_stats = sendonly.get_stats()
+    recvonly_stats = recvonly.get_stats()
+
+    sendonly.disconnect()
+    recvonly.disconnect()
+
+    # codec が無かったら StopIteration 例外が上がる
+    sendonly_codec_stats = next(s for s in sendonly_stats if s.get("type") == "codec")
+    assert sendonly_codec_stats["mimeType"] == "video/AV1"
+
+    # outbound-rtp が無かったら StopIteration 例外が上がる
+    outbound_rtp_stats = next(s for s in sendonly_stats if s.get("type") == "outbound-rtp")
+    assert outbound_rtp_stats["encoderImplementation"] == "libaom"
+    assert outbound_rtp_stats["bytesSent"] > 0
+    assert outbound_rtp_stats["packetsSent"] > 0
+
+    # codec が無かったら StopIteration 例外が上がる
+    recvonly_codec_stats = next(s for s in recvonly_stats if s.get("type") == "codec")
+    assert recvonly_codec_stats["mimeType"] == "video/AV1"
+
+    # outbound-rtp が無かったら StopIteration 例外が上がる
+    inbound_rtp_stats = next(s for s in recvonly_stats if s.get("type") == "inbound-rtp")
+    assert inbound_rtp_stats["decoderImplementation"] == "dav1d"
     assert inbound_rtp_stats["bytesReceived"] > 0
     assert inbound_rtp_stats["packetsReceived"] > 0
